Fix control request timeout issues
Turns out XUD can't handle being told to ack a control request and then
not returning from the handler in a timely fashion: we need to do our
work (update the mixer) *before* calling XUD_DoSetRequestStatus because
if we burn cycles afterwards and a new request comes in soon, it'll
fail. You'd think it would know how to NAK that until we're ready for
it, but nope.